fix(server): resolve login shell env in the daemon
The daemon inherited whatever environment its launcher had (launchd, a
minimal shell, a restart from a degraded session), so terminals fell
back to a bare /bin/bash with an incomplete PATH on macOS, and agent
processes saw the same broken environment. The desktop app's
inheritLoginShellEnv only ran in the Electron main process and never
helped a daemon started without it.

Move the login-shell-env implementation into @getpaseo/server (logger
is now injectable; the desktop package keeps a thin wrapper that passes
electron-log) and call it from the daemon worker bootstrap. An
inherited SHELL reflects the launcher rather than the user's login
shell, so prefer the account's login shell (os.userInfo().shell) when
one is registered.

Closes #2247
